Feat/inventory (#20)
* ci(api): run only minio and postgres for integration test

* docs(api): update README

* ci(api): add coverage code with codecov

* ci(api): fix codecov step

* test(api/product): cover not found case for delete, find one, update

* ci(codecov): try to upload report to codecov

* docs(README): update badges and prerequire

* test(api/category): update delete category integration test

* docs(README): update ci badge url

* style(api/category): rename integration test file

* feat(api/inventory): change inventory entities fields

* feat(api/utils): create setup gorm util func for integration test

* test(api/inventory): add suite for inventory integration test

* feat(api/inventory): add gorm repository implementation

+++ apps/api/src/product/domain/usecases/find_test.go
@@ -178,14 +178,14 @@ func (s *FindProductTestSuite) TestFindHappyCase() {
 		},
 		{
 			Queries: &domain.ProductQuery{
-				InventoryId: utils.GetDataTypeAddress(10),
+				InventoryId: utils.GetDataTypeAddress(uint(10)),
 			},
 			Expected:    []*domain.Product{},
 			Description: "Query By Inventory Id Not Found",
 		},
 		{
 			Queries: &domain.ProductQuery{
-				InventoryId: utils.GetDataTypeAddress(1),
+				InventoryId: utils.GetDataTypeAddress(uint(1)),
 			},
 			Expected:    []*domain.Product{&initData[0], &initData[1]},
 			Description: "Query By Inventory Id Happy Case",
@@ -197,7 +197,7 @@ func (s *FindProductTestSuite) TestFindHappyCase() {
 					PageSize: utils.GetDataTypeAddress(uint(2)),
 					Q:        utils.GetDataTypeAddress("free"),
 				},
-				InventoryId: utils.GetDataTypeAddress(2),
+				InventoryId: utils.GetDataTypeAddress(uint(2)),
 			},
 			Expected:    []*domain.Product{&initData[2]},
 			Description: "Query By Inventory Id All Params",
